chore: moved theme tests to hooks folder.
